improve handling of backends and registration of plotfuncs
